cancel download

+++ app/actions/download.js
@@ -13,6 +13,10 @@ let exercisesList = [];
 let exerciseLengthList = [];
 let exerciseIdList = [];
 let task;
+let introExerciseTask = [];
+let introImageTask = [];
+let advanceImageTask = [];
+let advanceExerciseTask = [];
 
 const saveEpisodeToDatabase = (
   episodeTitle,
@@ -29,6 +33,7 @@ const saveEpisodeToDatabase = (
   endWT,
 ) => {
   realm.write(() => {
+    console.log('SAVED EPISODES');
     realm.create('SavedEpisodes', {
       id: episodeId,
       title: episodeTitle,
@@ -73,7 +78,6 @@ const startDownload = (
   let episodeExerciseSize = 0;
   let episodeSizeReceived = 0;
   let count = 0;
-
   const totalVideoSizeInBytes = (videoSize * 1000 * 1000);
   dispatch({
     type: ACTION_DOWNLOAD,
@@ -93,147 +97,147 @@ const startDownload = (
         type: ACTION_DOWNLOAD_PROGRESS,
         payload: (received / totalVideoSizeInBytes),
       });
-    })
-    .then(() => {
-      if (exercisesList.length === 0) {
-        saveEpisodeToDatabase(
-          episodeTitle,
-          episodeId,
-          category,
-          description,
-          video,
-          totalTime,
-          workoutTime,
-          videoSize,
-          episodeIndex,
-          seriesIndex,
-          startWT,
-          endWT,
-          dispatch,
-        );
+    });
+  task.then(() => {
+    if (exercisesList.length === 0) {
+      saveEpisodeToDatabase(
+        episodeTitle,
+        episodeId,
+        category,
+        description,
+        video,
+        totalTime,
+        workoutTime,
+        videoSize,
+        episodeIndex,
+        seriesIndex,
+        startWT,
+        endWT,
+        dispatch,
+      );
+      return;
+    }
+    exercisesList.map((exercise, i) => {
+      const {
+        cmsTitle, image, title, advanced, id, visible, episodeExerciseTitle, exerciseSize, index,
+      } = exercise;
+      const formattedExerciseName = cmsTitle.replace(/\s+/g, '');
+      realm.write(() => {
+        realm.create('SavedExercises', {
+          id,
+          title,
+          cmsTitle,
+          visible,
+          episodeExerciseTitle,
+          index,
+          advanced: advanced === undefined ? false : true,
+        });
+      });
+      if (exercise.video === '') {
+        introImageTask[i] = RNFetchBlob
+          .config({
+            path: `${dirs.DocumentDir}/AST/introImages/${formattedExerciseName}.png`,
+          })
+          .fetch('GET', `${image}`, {
+          });
+        introImageTask[i].then(() => {
+          advanceImageTask = RNFetchBlob
+            .config({
+              path: `${dirs.DocumentDir}/AST/advanceImages/${formattedExerciseName}.png`,
+            })
+            .fetch('GET', advanced === undefined ? image : advanced.image, {
+            });
+          advanceImageTask.then(() => {
+            episodeExerciseSize = count === 0 ? (episodeSizeReceived + exerciseSize) : (episodeExerciseSize + exerciseSize);
+            count += 1;
+            dispatch({
+              type: ACTION_DOWNLOAD_PROGRESS,
+              payload: (episodeExerciseSize / totalVideoSizeInBytes),
+            });
+            if (i === (exercisesList.length - 1)) {
+              saveEpisodeToDatabase(
+                episodeTitle,
+                episodeId,
+                category,
+                description,
+                video,
+                totalTime,
+                workoutTime,
+                videoSize,
+                episodeIndex,
+                seriesIndex,
+                startWT,
+                endWT,
+                dispatch,
+              );
+            }
+          })
+            .catch(error => console.log(error));
+        })
+          .catch(error => console.log(error));
         return;
       }
-      exercisesList.map((exercise, i) => {
-        const {
-          cmsTitle, title, image, advanced, id, visible, episodeExerciseTitle, exerciseSize, index,
-        } = exercise;
-        const formattedExerciseName = cmsTitle.replace(/\s+/g, '');
-        realm.write(() => {
-          realm.create('SavedExercises', {
-            id,
-            title,
-            cmsTitle,
-            visible,
-            episodeExerciseTitle,
-            index,
-            advanced: advanced === undefined ? false : true,
-          });
+      introImageTask[i] = RNFetchBlob
+        .config({
+          path: `${dirs.DocumentDir}/AST/introImages/${formattedExerciseName}.png`,
+        })
+        .fetch('GET', `${image}`, {
         });
-        if (exercise.video === '') {
-          RNFetchBlob
-            .config({
-              path: `${dirs.DocumentDir}/AST/introImages/${formattedExerciseName}.png`,
-            })
-            .fetch('GET', `${image}`, {
-            })
-            .then(() => {
-              RNFetchBlob
-                .config({
-                  path: `${dirs.DocumentDir}/AST/advanceImages/${formattedExerciseName}.png`,
-                })
-                .fetch('GET', advanced === undefined ? image : advanced.image, {
-                })
-                .then(() => {
-                  episodeExerciseSize = count === 0 ? (episodeSizeReceived + exerciseSize) : (episodeExerciseSize + exerciseSize);
-                  count += 1;
-                  dispatch({
-                    type: ACTION_DOWNLOAD_PROGRESS,
-                    payload: (episodeExerciseSize / totalVideoSizeInBytes),
-                  });
-                  if (i === (exercisesList.length - 1)) {
-                    saveEpisodeToDatabase(
-                      episodeTitle,
-                      episodeId,
-                      category,
-                      description,
-                      video,
-                      totalTime,
-                      workoutTime,
-                      videoSize,
-                      episodeIndex,
-                      seriesIndex,
-                      startWT,
-                      endWT,
-                      dispatch,
-                    );
-                  }
-                })
-                .catch(error => console.log(error));
-            })
-            .catch(error => console.log(error));
-          return;
-        }
-        RNFetchBlob
+      introImageTask[i].then(() => {
+        introExerciseTask = RNFetchBlob
           .config({
             path: `${dirs.DocumentDir}/AST/introExercises/${formattedExerciseName}.mp4`,
           })
           .fetch('GET', `${exercise.video}`, {
-          })
-          .then(() => {
-            RNFetchBlob
+          });
+        introExerciseTask.then(() => {
+          advanceExerciseTask = RNFetchBlob
+            .config({
+              path: `${dirs.DocumentDir}/AST/advanceExercises/${formattedExerciseName}.mp4`,
+            })
+            .fetch('GET', advanced === undefined ? exercise.video : advanced.video, {
+            });
+          advanceExerciseTask.then(() => {
+            advanceImageTask = RNFetchBlob
               .config({
-                path: `${dirs.DocumentDir}/AST/introImages/${formattedExerciseName}.png`,
+                path: `${dirs.DocumentDir}/AST/advanceImages/${formattedExerciseName}.png`,
               })
-              .fetch('GET', `${image}`, {
-              })
-              .then(() => {
-                RNFetchBlob
-                  .config({
-                    path: `${dirs.DocumentDir}/AST/advanceExercises/${formattedExerciseName}.mp4`,
-                  })
-                  .fetch('GET', advanced === undefined ? exercise.video : advanced.video, {
-                  })
-                  .then(() => {
-                    RNFetchBlob
-                      .config({
-                        path: `${dirs.DocumentDir}/AST/advanceImages/${formattedExerciseName}.png`,
-                      })
-                      .fetch('GET', advanced === undefined ? image : advanced.image, {
-                      })
-                      .then(() => {
-                        episodeExerciseSize += exerciseSize;
-                        count += 1;
-                        dispatch({
-                          type: ACTION_DOWNLOAD_PROGRESS,
-                          payload: (episodeExerciseSize / totalVideoSizeInBytes),
-                        });
-                        if (i === (exercisesList.length - 1)) {
-                          saveEpisodeToDatabase(
-                            episodeTitle,
-                            episodeId,
-                            category,
-                            description,
-                            video,
-                            totalTime,
-                            workoutTime,
-                            videoSize,
-                            episodeIndex,
-                            seriesIndex,
-                            startWT,
-                            endWT,
-                            dispatch,
-                          );
-                        }
-                      })
-                      .catch(error => console.log(error));
-                  })
-                  .catch(error => console.log(error));
-              })
+              .fetch('GET', advanced === undefined ? image : advanced.image, {
+              });
+            advanceImageTask.then(() => {
+              episodeExerciseSize += exerciseSize;
+              count += 1;
+              dispatch({
+                type: ACTION_DOWNLOAD_PROGRESS,
+                payload: (episodeExerciseSize / totalVideoSizeInBytes),
+              });
+              if (i === (exercisesList.length - 1)) {
+                saveEpisodeToDatabase(
+                  episodeTitle,
+                  episodeId,
+                  category,
+                  description,
+                  video,
+                  totalTime,
+                  workoutTime,
+                  videoSize,
+                  episodeIndex,
+                  seriesIndex,
+                  startWT,
+                  endWT,
+                  dispatch,
+                );
+              }
+            })
               .catch(error => console.log(error));
           })
+            .catch(error => console.log(error));
+        })
           .catch(error => console.log(error));
-      });
-    })
+      })
+        .catch(error => console.log(error));
+    });
+  })
     .catch(error => console.log(error));
 };
 
@@ -245,14 +249,6 @@ const startDownload = (
 //     payload: (dataReceived / totalVideoSizeInBytes),
 //   });
 // })
-
-export const stopDownload = () => {
-  task.cancel(err => console.log(err));
-  return {
-    type: ACTION_DOWNLOAD_CANCEL,
-  };
-};
-
 
 export const downloadEpisode = ({
   exercises,
@@ -269,6 +265,13 @@ export const downloadEpisode = ({
   startWT,
   endWT,
 }) => {
+  advanceImageTask = [];
+  introExerciseTask = [];
+  advanceExerciseTask = [];
+  introImageTask = [];
+  exercisesList = [];
+  exerciseLengthList = [];
+  exerciseIdList = [];
   return async (dispatch) => {
     if (exercises !== undefined) {
       await exercises.map(async (value, i) => {
@@ -277,6 +280,10 @@ export const downloadEpisode = ({
         } = value;
         exerciseLengthList.push(length);
         exerciseIdList.push(uid);
+        introExerciseTask.push(i);
+        introImageTask.push(i);
+        advanceImageTask.push(i);
+        advanceExerciseTask.push(i);
         await firebase.database().ref(`exercises/${uid}`).on('value', (snapShot) => {
           const exercise = { ...snapShot.val(), id: uid, visible, episodeExerciseTitle, index: i };
           exercisesList.push(exercise);
@@ -317,13 +324,14 @@ export const downloadEpisode = ({
   };
 };
 
-export const deleteEpisodeList = fileName => (dispatch) => {
-  dispatch({
-    type: ACTION_DELETE_EPISODE,
-    payload: false,
-  });
+const deleteEpisode = (fileName, dispatch) => {
   const { dirs } = RNFetchBlob.fs;
   const episodeDetail = Array.from(realm.objects('SavedEpisodes').filtered(`title="${fileName}"`));
+  console.log(fileName);
+  console.log(episodeDetail.length);
+  if (episodeDetail.length === 0) {
+    return;
+  }
   const exerciseIdLists = Array.from(episodeDetail[0].exerciseIdList);
   const allEpisodes = Array.from(realm.objects('SavedEpisodes'));
   const formattedFileName = fileName.replace(/ /g, '_');
@@ -340,6 +348,9 @@ export const deleteEpisodeList = fileName => (dispatch) => {
         });
         if (count < 2) {
           const exerciseDetail = realm.objects('SavedExercises').filtered(`id="${value}"`);
+          if ((Array.from(exerciseDetail)).length === 0) {
+            return;
+          }
           const exerciseTitle = Array.from(exerciseDetail)[0].cmsTitle;
           const formattedExerciseTitle = exerciseTitle.replace(/\s+/g, '');
           RNFetchBlob.fs.unlink(`${dirs.DocumentDir}/AST/episodes/${formattedExerciseTitle}.mp4`)
@@ -374,4 +385,47 @@ export const deleteEpisodeList = fileName => (dispatch) => {
     .catch((err) => {
       console.log(err);
     });
+};
+
+export const deleteEpisodeList = fileName => (dispatch) => {
+  dispatch({
+    type: ACTION_DELETE_EPISODE,
+    payload: false,
+  });
+  deleteEpisode(fileName, dispatch);
+};
+
+export const stopDownload = title => async (dispatch) => {
+  task.cancel(err => console.log(err));
+  console.log(introExerciseTask.length);
+  console.log(introExerciseTask);
+  console.log(introImageTask);
+  console.log(advanceExerciseTask);
+  console.log(advanceImageTask);
+  await introImageTask.map((value, i) => {
+    console.log(i);
+    if (introImageTask[i] !== i && introImageTask[i] !== undefined) {
+      console.log(i);
+      const intImgTask = introImageTask[i];
+      intImgTask.cancel(err => console.log(err));
+    }
+    if (introExerciseTask[i] !== i && introExerciseTask[i] !== undefined) {
+      const intExTask = introExerciseTask[i];
+      intExTask.cancel(err => console.log(err));
+    }
+    if (advanceExerciseTask[i] !== i && advanceExerciseTask[i] !== undefined) {
+      const advExTask = advanceExerciseTask[i];
+      advExTask.cancel(err => console.log(err));
+    }
+    if (advanceImageTask[i] !== i && advanceImageTask[i] !== undefined) {
+      const advImgTask = advanceImageTask[i];
+      advImgTask.cancel(err => console.log(err));
+    }
+  });
+  dispatch({
+    type: ACTION_DOWNLOAD_CANCEL,
+  });
+  setTimeout(() => {
+    deleteEpisode(title, dispatch);
+  }, 5000);
 };
